+++ mountpoint-s3-crt-sys/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+* Update to latest CRT dependencies
+
 ## v0.7.0 (April 10, 2024)
 
 * Update to latest CRT dependencies